requirements.txt
langchain_community==0.3.10
langchain_openai==0.2.12
openai==1.57.1
python-dotenv==1.0.1
